+++ cpp/src/proto/substrait_to_velox_expr.cc
@@ -1,10 +1,21 @@
-#include "substrait_to_velox_expr.h"
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-#include <arrow/array/array_primitive.h>
-#include <arrow/array/data.h>
-#include <arrow/array/util.h>
-#include <arrow/record_batch.h>
-#include <arrow/type_fwd.h>
+#include "substrait_to_velox_expr.h"
 
 namespace substrait = io::substrait;
 using namespace facebook::velox;
@@ -17,6 +28,100 @@ SubstraitVeloxExprConverter::SubstraitVeloxExprConverter(
     const std::unordered_map<uint64_t, std::string>& functions_map) {
   sub_parser_ = sub_parser;
   functions_map_ = functions_map;
+}
+
+std::shared_ptr<const core::FieldAccessTypedExpr>
+SubstraitVeloxExprConverter::toVeloxExpr(const substrait::FieldReference& sfield,
+                                         const int32_t& input_plan_node_id) {
+  switch (sfield.reference_type_case()) {
+    case substrait::FieldReference::ReferenceTypeCase::kDirectReference: {
+      auto dref = sfield.direct_reference();
+      int32_t col_idx = parseReferenceSegment(dref);
+      auto field_name = sub_parser_->makeNodeName(input_plan_node_id, col_idx);
+      // FIXME: How to get the input types?
+      return std::make_shared<const core::FieldAccessTypedExpr>(DOUBLE(), field_name);
+      break;
+    }
+    case substrait::FieldReference::ReferenceTypeCase::kMaskedReference: {
+      throw new std::runtime_error("not supported");
+      break;
+    }
+    default:
+      throw new std::runtime_error("not supported");
+      break;
+  }
+}
+
+std::shared_ptr<const core::ITypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
+    const substrait::Expression::ScalarFunction& sfunc,
+    const int32_t& input_plan_node_id) {
+  std::vector<std::shared_ptr<const core::ITypedExpr>> params;
+  for (auto& sarg : sfunc.args()) {
+    auto expr = toVeloxExpr(sarg, input_plan_node_id);
+    params.push_back(expr);
+  }
+  auto function_id = sfunc.id().id();
+  auto function_name = sub_parser_->findFunction(functions_map_, function_id);
+  auto velox_function = sub_parser_->substrait_velox_function_map[function_name];
+  auto sub_type = sub_parser_->parseType(sfunc.output_type());
+  auto velox_type = getVeloxType(sub_type->type);
+  return std::make_shared<const core::CallTypedExpr>(velox_type, std::move(params),
+                                                     velox_function);
+}
+
+std::shared_ptr<const core::ConstantTypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
+    const io::substrait::Expression::Literal& slit) {
+  switch (slit.literal_type_case()) {
+    case substrait::Expression_Literal::LiteralTypeCase::kFp64: {
+      double val = slit.fp64();
+      return std::make_shared<core::ConstantTypedExpr>(val);
+      break;
+    }
+    case substrait::Expression_Literal::LiteralTypeCase::kBoolean: {
+      bool val = slit.boolean();
+      throw new std::runtime_error("Type is not supported.");
+      break;
+    }
+    default:
+      throw new std::runtime_error("Type is not supported.");
+      break;
+  }
+}
+
+std::shared_ptr<const core::ITypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
+    const substrait::Expression& sexpr, const int32_t& input_plan_node_id) {
+  std::shared_ptr<const core::ITypedExpr> velox_expr;
+  switch (sexpr.rex_type_case()) {
+    case substrait::Expression::RexTypeCase::kLiteral: {
+      auto slit = sexpr.literal();
+      velox_expr = toVeloxExpr(slit);
+      break;
+    }
+    case substrait::Expression::RexTypeCase::kScalarFunction: {
+      auto sfunc = sexpr.scalar_function();
+      velox_expr = toVeloxExpr(sfunc, input_plan_node_id);
+      break;
+    }
+    case substrait::Expression::RexTypeCase::kSelection: {
+      auto sel = sexpr.selection();
+      velox_expr = toVeloxExpr(sel, input_plan_node_id);
+      break;
+    }
+    default:
+      throw new std::runtime_error("Expression not supported");
+      break;
+  }
+  return velox_expr;
+}
+
+TypePtr SubstraitVeloxExprConverter::getVeloxType(std::string type_name) {
+  if (type_name == "BOOL") {
+    return BOOLEAN();
+  } else if (type_name == "FP64") {
+    return DOUBLE();
+  } else {
+    throw std::runtime_error("Type name is not supported");
+  }
 }
 
 int32_t SubstraitVeloxExprConverter::parseReferenceSegment(
@@ -34,10 +139,10 @@ int32_t SubstraitVeloxExprConverter::parseReferenceSegment(
   }
 }
 
+// This class is used by Filter PushDown.
 class SubstraitVeloxExprConverter::FilterInfo {
  public:
   FilterInfo() {}
-
   void setLeft(double left, bool isExclusive) {
     left_ = left;
     left_exclusive_ = isExclusive;
@@ -53,7 +158,7 @@ class SubstraitVeloxExprConverter::FilterInfo {
     }
   }
   void forbidsNull() {
-    allow_null_ = false;
+    null_allowed_ = false;
     if (!is_initialized_) {
       is_initialized_ = true;
     }
@@ -62,7 +167,7 @@ class SubstraitVeloxExprConverter::FilterInfo {
 
   std::optional<double> left_ = std::nullopt;
   std::optional<double> right_ = std::nullopt;
-  bool allow_null_ = true;
+  bool null_allowed_ = true;
   bool left_exclusive_ = false;
   bool right_exclusive_ = false;
 
@@ -159,7 +264,7 @@ hive::SubfieldFilters SubstraitVeloxExprConverter::toVeloxFilter(
         right_bound = filter_info->right_.value();
         right_exclusive = filter_info->right_exclusive_;
       }
-      bool null_allowed = filter_info->allow_null_;
+      bool null_allowed = filter_info->null_allowed_;
       filters[common::Subfield(input_name_list[idx])] =
           std::make_unique<common::DoubleRange>(
               left_bound, left_unbounded, left_exclusive, right_bound, right_unbounded,
@@ -167,88 +272,4 @@ hive::SubfieldFilters SubstraitVeloxExprConverter::toVeloxFilter(
     }
   }
   return filters;
-}
-
-std::shared_ptr<const core::FieldAccessTypedExpr>
-SubstraitVeloxExprConverter::toVeloxExpr(const substrait::FieldReference& sfield,
-                                         const int32_t& input_plan_node_id) {
-  switch (sfield.reference_type_case()) {
-    case substrait::FieldReference::ReferenceTypeCase::kDirectReference: {
-      auto dref = sfield.direct_reference();
-      int32_t col_idx = parseReferenceSegment(dref);
-      auto field_name = sub_parser_->makeNodeName(input_plan_node_id, col_idx);
-      // FIXME: How to get the input types?
-      return std::make_shared<const core::FieldAccessTypedExpr>(DOUBLE(), field_name);
-      break;
-    }
-    case substrait::FieldReference::ReferenceTypeCase::kMaskedReference: {
-      throw new std::runtime_error("not supported");
-      break;
-    }
-    default:
-      throw new std::runtime_error("not supported");
-      break;
-  }
-}
-
-std::shared_ptr<const core::ITypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
-    const substrait::Expression::ScalarFunction& sfunc,
-    const int32_t& input_plan_node_id) {
-  std::vector<std::shared_ptr<const core::ITypedExpr>> params;
-  for (auto& sarg : sfunc.args()) {
-    auto expr = toVeloxExpr(sarg, input_plan_node_id);
-    params.push_back(expr);
-  }
-  auto function_id = sfunc.id().id();
-  auto function_name = sub_parser_->findFunction(functions_map_, function_id);
-  auto velox_function = sub_parser_->substrait_velox_function_map[function_name];
-  auto sub_type = sub_parser_->parseType(sfunc.output_type());
-  auto velox_type = sub_parser_->getVeloxType(sub_type->type);
-  return std::make_shared<const core::CallTypedExpr>(velox_type, std::move(params),
-                                                     velox_function);
-}
-
-std::shared_ptr<const core::ConstantTypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
-    const io::substrait::Expression::Literal& slit) {
-  switch (slit.literal_type_case()) {
-    case substrait::Expression_Literal::LiteralTypeCase::kFp64: {
-      double val = slit.fp64();
-      return std::make_shared<core::ConstantTypedExpr>(val);
-      break;
-    }
-    case substrait::Expression_Literal::LiteralTypeCase::kBoolean: {
-      bool val = slit.boolean();
-      throw new std::runtime_error("Type is not supported.");
-      break;
-    }
-    default:
-      throw new std::runtime_error("Type is not supported.");
-      break;
-  }
-}
-
-std::shared_ptr<const core::ITypedExpr> SubstraitVeloxExprConverter::toVeloxExpr(
-    const substrait::Expression& sexpr, const int32_t& input_plan_node_id) {
-  std::shared_ptr<const core::ITypedExpr> velox_expr;
-  switch (sexpr.rex_type_case()) {
-    case substrait::Expression::RexTypeCase::kLiteral: {
-      auto slit = sexpr.literal();
-      velox_expr = toVeloxExpr(slit);
-      break;
-    }
-    case substrait::Expression::RexTypeCase::kScalarFunction: {
-      auto sfunc = sexpr.scalar_function();
-      velox_expr = toVeloxExpr(sfunc, input_plan_node_id);
-      break;
-    }
-    case substrait::Expression::RexTypeCase::kSelection: {
-      auto sel = sexpr.selection();
-      velox_expr = toVeloxExpr(sel, input_plan_node_id);
-      break;
-    }
-    default:
-      throw new std::runtime_error("Expression not supported");
-      break;
-  }
-  return velox_expr;
 }
