+++ cpp/src/proto/substrait_utils.h
@@ -29,32 +29,9 @@
 #include "selection.pb.h"
 #include "type.pb.h"
 #include "type_expressions.pb.h"
-#include "velox/buffer/Buffer.h"
-#include "velox/common/caching/DataCache.h"
-#include "velox/common/file/FileSystems.h"
-#include "velox/connectors/hive/FileHandle.h"
-#include "velox/connectors/hive/HiveConnector.h"
-#include "velox/connectors/hive/HiveConnectorSplit.h"
-#include "velox/core/Expressions.h"
-#include "velox/core/ITypedExpr.h"
-#include "velox/core/PlanNode.h"
-#include "velox/dwio/common/Options.h"
-#include "velox/dwio/common/ScanSpec.h"
-#include "velox/dwio/dwrf/common/CachedBufferedInput.h"
-#include "velox/dwio/dwrf/reader/DwrfReader.h"
-#include "velox/dwio/dwrf/writer/Writer.h"
-#include "velox/exec/Operator.h"
-#include "velox/exec/OperatorUtils.h"
-#include "velox/exec/tests/utils/Cursor.h"
-#include "velox/expression/Expr.h"
-#include "velox/functions/prestosql/aggregates/SumAggregate.h"
-#include "velox/functions/prestosql/registration/RegistrationFunctions.h"
-#include "velox/type/Filter.h"
-#include "velox/type/Subfield.h"
 
-using namespace facebook::velox;
-using namespace facebook::velox::exec;
-
+// This class contains some common funcitons used to parse Substrait components, and
+// convert it to recognizable representations.
 class SubstraitParser {
  public:
   SubstraitParser();
@@ -62,16 +39,15 @@ class SubstraitParser {
     std::string type;
     std::string name;
     bool nullable;
-    SubstraitType(const std::string& t, const std::string& n, const bool& nul) {
+    SubstraitType(const std::string& t, const std::string& n, const bool& null) {
       type = t;
       name = n;
-      nullable = nul;
+      nullable = null;
     }
   };
   std::vector<std::shared_ptr<SubstraitParser::SubstraitType>> parseNamedStruct(
       const io::substrait::Type::NamedStruct& named_struct);
   std::shared_ptr<SubstraitType> parseType(const io::substrait::Type& stype);
-  TypePtr getVeloxType(std::string type_name);
   std::vector<std::string> makeNames(const std::string& prefix, int size);
   std::string makeNodeName(int node_id, int col_idx);
   std::string findFunction(const std::unordered_map<uint64_t, std::string>& functions_map,
@@ -79,10 +55,4 @@ class SubstraitParser {
   // Used for mapping Substrait function key word into Velox functions.
   std::unordered_map<std::string, std::string> substrait_velox_function_map = {
       {"MULTIPLY", "multiply"}, {"SUM", "sum"}};
-};
-
-class VeloxInitializer {
- public:
-  VeloxInitializer();
-  void Init();
 };
